+++ ModelDb/MailboxData.cs
@@ -28,8 +28,8 @@
         public Int32 ItemCode { get; set; } = 0;
         public Int32 Amount { get; set; } = 0;
         public Int32 EnchanceCount { get; set; } = 0;
-        public Int32 Attack { get; set; } = 0;
-        public Int32 Defence { get; set; } = 0;
-        public Int32 Magic { get; set; } = 0;
+        public decimal Attack { get; set; } = 0;
+        public decimal Defence { get; set; } = 0;
+        public decimal Magic { get; set; } = 0;
     }
 }
